add batching/parallelization to oracle reads
fixes #1100
